Finish script
